chore : 클래스명 변경
